Renamed and refactored the LookAndFeel classes. The LookAndFeel class is now an abstract base-class, and there are 3 concrete subclasses you can use. These are LookAndFeel_V1 (replaces OldSchoolLookAndFeel), LookAndFeel_V2 (replaces what was the normal LookAndFeel class) and LookAndFeel_V3 (this is a new style, taken from the Introjucer's L+F).

+++ modules/juce_audio_formats/codecs/juce_FlacAudioFormat.cpp
@@ -380,7 +380,7 @@ public:
             samplesToWrite = const_cast<const int**> (channels.getData());
         }
 
-        return FLAC__stream_encoder_process (encoder, (const FLAC__int32**) samplesToWrite, (size_t) numSamples) != 0;
+        return FLAC__stream_encoder_process (encoder, (const FLAC__int32**) samplesToWrite, (unsigned) numSamples) != 0;
     }
 
     bool writeData (const void* const data, const int size) const
