+++ modules/juce_audio_formats/codecs/juce_OggVorbisAudioFormat.cpp
@@ -186,7 +186,7 @@ public:
                 {
                     float** dataIn = nullptr;
 
-                    const int samps = OggVorbisNamespace::ov_read_float (&ovFile, &dataIn, numToRead, &bitStream);
+                    const long samps = OggVorbisNamespace::ov_read_float (&ovFile, &dataIn, numToRead, &bitStream);
                     if (samps <= 0)
                         break;
